+++ src/configuration.cpp
@@ -157,7 +157,6 @@ bool configure(int argc, char **argv)
 		if (vm.count("test-lyrics-fetchers"))
 		{
 			std::vector<std::tuple<std::string, std::string, std::string>> fetcher_data = {
-				std::make_tuple("genius", "rihanna", "umbrella"),
 				std::make_tuple("justsomelyrics", "rihanna", "umbrella"),
 				std::make_tuple("jahlyrics", "sean kingston", "dry your eyes"),
 				std::make_tuple("plyrics", "rihanna", "umbrella"),